requirements.txt
sentence-transformers
pdfplumber
tf-keras
polars-u64-idx